backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-multipart
pydantic
python-dotenv
requests
python-dotenv